Added support for specifying the dead letter routing key.
The x-dead-letter-routing-key queue argument is used in conjunction with the
x-dead-letter-exchange argument. When specified, it causes messages to be sent
to the deadletter exchange with the specified routing key rather than the one
they were originally published with. See http://www.rabbitmq.com/dlx.html
for more information.

Also added the WithDeadLetterExchange() method to IConsumeConfigurationContext
since I overlooked that in my prior commit.